+++ NEWS.md
@@ -1,4 +1,4 @@
-# attachment 0.2.4.900x
+# attachment 0.2.5
 ## Major changes
 
 * add `create_renv_for_dev()` and `create_renv_for_prod()` function to create `renv.lock` file based on development project (@VincentGuyader and @statnmap).